fix msp crc read/flush race condition
We were reading in data packets up to, but not including the checksum
since we didn't use it anyways. Instead of reading in the checksum,
we would flush the serial port assuming that the checksum would be
flushed. However, if the checksum byte had not arrived yet, flushing
the buffer would not clear the checksum byte when it did arrive. That
is, we had a race condition between clearing the buffer and the checksum
byte being read. If the buffer flush came first, future packets would
be offset by 1 and major headaches would ensue.