redo drone_comm for MSP-only communication
Moving codebase away from PWM based communication
PWM communications require continuous timing measurements and are a
holdover from past days in the RC hobby. Digital hadrware (like the
Raspberry Pi that this codebase is running on) is not a good fit for
continuous time measurements.

MultiWii Serial Protocol communications can be carried over USB connections as
digital, serial communications and are a much better fit for the digital
hardware.